refactor: remove unused V1 diagram implementation
