+++ operate/client/src/App/Processes/MigrationView/TopPanel/Diagrams/v2/SourceDiagram.tsx
@@ -11,7 +11,7 @@ import {Header} from '../Header';
 import {DiagramWrapper} from '../styled';
 import {observer} from 'mobx-react';
 import {DiagramShell} from 'modules/components/DiagramShell';
-import {Diagram} from 'modules/components/Diagram/v2';
+import {Diagram} from 'modules/components/Diagram';
 import {processInstanceMigrationStore} from 'modules/stores/processInstanceMigration';
 import {diagramOverlaysStore} from 'modules/stores/diagramOverlays';
 import {StateOverlay} from 'modules/components/StateOverlay';
