+++ operate/client/src/modules/components/Diagram/index.test.tsx
@@ -11,6 +11,9 @@ import {observer} from 'mobx-react';
 import {diagramOverlaysStore} from 'modules/stores/diagramOverlays';
 import {createPortal} from 'react-dom';
 import {Diagram} from './index';
+import {MemoryRouter, Route, Routes} from 'react-router-dom';
+import {Paths} from 'modules/Routes';
+import {ProcessDefinitionKeyContext} from 'App/Processes/ListView/processDefinitionKeyContext';
 import {QueryClientProvider} from '@tanstack/react-query';
 import {getMockQueryClient} from 'modules/react-query/mockQueryClient';
 
@@ -21,11 +24,17 @@ const Overlay: React.FC<{container: HTMLElement; data: React.ReactNode}> =
     return createPortal(<div>{data}</div>, container);
   });
 
-const Wrapper: React.FC<{children?: React.ReactNode}> = ({children}) => {
+const Wrapper = ({children}: {children?: React.ReactNode}) => {
   return (
-    <QueryClientProvider client={getMockQueryClient()}>
-      {children}
-    </QueryClientProvider>
+    <MemoryRouter initialEntries={[Paths.processInstance('1')]}>
+      <ProcessDefinitionKeyContext.Provider value="123">
+        <QueryClientProvider client={getMockQueryClient()}>
+          <Routes>
+            <Route path={Paths.processInstance()} element={children} />
+          </Routes>
+        </QueryClientProvider>
+      </ProcessDefinitionKeyContext.Provider>
+    </MemoryRouter>
   );
 };
